Move Game class into its own set of files
