+++ src/game_main.hpp
@@ -16,23 +16,10 @@
 
 #pragma once
 
-#include "base/warnings.hpp"
 #include "common/command_line_options.hpp"
-
-RIGEL_DISABLE_WARNINGS
-#include <SDL_video.h>
-RIGEL_RESTORE_WARNINGS
-
-#include <memory>
 
 
 namespace rigel {
-
-class Game;
-class UserProfile;
-
-// Regular game main
-
 
 /** Initialize platform, create game, run main loop
  *
@@ -45,43 +32,5 @@ class UserProfile;
  * the options menu.
  */
 void gameMain(const CommandLineOptions& options);
-
-
-// Lower-level API for integration into callback-based frameworks
-
-struct GameDeleter {
-  void operator()(Game* pGame);
-};
-
-/** Create a Game
- *
- * This creates a Game for use with runOneFrame. It requires the user profile
- * to have a valid game path set, otherwise it will fail with an exception.
- * It also needs a valid OpenGL window to already be open, and ImGui must have
- * been initialized.
- *
- * This is meant for environments where Rigel cannot own the main loop, like
- * when running in a browser via Webassembly.
- *
- * TODO: Move the platform initialization code to a place where it can be
- * reused more easily, independently of gameMain().
- */
-std::unique_ptr<Game, GameDeleter> createGame(
-  SDL_Window* pWindow,
-  UserProfile* pUserProfile,
-  const CommandLineOptions& commandLineOptions);
-
-/** Run one frame of the game
- *
- * Use this with the Game object returned from createGame. Should be called
- * once every frame. This is meant for callback-based environments, where the
- * main loop is outside of Rigel's control, like Webassembly.
- *
- * At the moment, this does not allow for game restarts as required when
- * choosing a different game path. But since createGame() doesn't implement
- * the game path browser and requires a valid game path to be pre-configured,
- * this seems like an ok limitation.
- */
-void runOneFrame(Game* pGame);
 
 }
